% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{clean_cadaster}
\alias{clean_cadaster}
\title{Clean the cadaster numbers}
\usage{
clean_cadaster(df_bp)
}
\arguments{
\item{df_bp}{An expanded data frame in wide format consisting of publications}
}
\description{
Clean cadaster numbers with regular expressions. Example: For a given publication
\code{c("A2000, A2001, Vers-Nr. 20003, 20004")} becomes \code{c("A2000, A2001")}.
}
