% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{clean_df_bp}
\alias{clean_df_bp}
\title{Clean raw publication data frame}
\usage{
clean_df_bp(df_bp)
}
\arguments{
\item{df_bp}{An expanded data frame in wide format consisting of publications}
}
\description{
This function cleans that raw publication data that was already processed to
wide data frame. Its main purpose is to clean and unnests the cadaster numbers.
}
