% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGD_gpkg.R
\name{create_gpkg}
\alias{create_gpkg}
\title{Create or Update a GeoPackage with Spatial Building Publication Data}
\usage{
create_gpkg(gpkg_file, df_bp, sf_liegenschaften)
}
\arguments{
\item{gpkg_file}{The name of the .gpkg-file. For example \code{baupub.gpkg}.}

\item{df_bp}{A data frame containing building permit information including}

\item{sf_liegenschaften}{A shapefile containing the Liegenschaften-data
geometry and metadata.}
}
\value{
Writes data to a GeoPackage file. Returns nothing.
}
\description{
Writes filtered and spatially prepared building permit data to a GeoPackage
file. If the file does not exist, it creates a new one. If it exists, it
appends only the new entries based on project IDs.
}
\examples{
\dontrun{
create_gpkg("baupub.gpkg", df_bp, sf_liegenschaften)
}
}
