% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_geo_output.R
\name{create_map}
\alias{create_map}
\title{Create an Interactive Leaflet Map of Building Permits}
\usage{
create_map(
  sf_bp_geo,
  start_date = as.character(Sys.Date() - 20),
  end_date = as.character(Sys.Date()),
  carto_basemap_api_key
)
}
\arguments{
\item{sf_bp_geo}{An \code{sf} object containing spatial building permit data,
including attributes such as \code{publicationNumber}, \code{entryDeadline},
\code{address}, and \code{url}.}

\item{start_date}{A character date in the format "yyyy-mm-dd" representing the
first day for which data should be included.
first day for which data should be included.}

\item{end_date}{A character date in the format "yyyy-mm-dd" representing the
the last day for which data should be included.}

\item{carto_basemap_api_key}{An api-key such that the background map is displayed
correctly. The key can be requested \href{https://carto.com/basemaps/apikey/}{here}
for free.}
}
\value{
A Leaflet map widget displaying the building permit polygons and
associated information.
}
\description{
Generates an interactive Leaflet map from spatial building permit data.
Each polygon includes a popup with details such as publication number,
publication date, entry deadline, address, and overlapping permits. The
default observation period is set to 20 days, as this is the timeframe during
which objections to the project can be submitted.
}
\examples{
\dontrun{
# creating a map of the last 20 days
create_map(sf_bp_geo)

# customize period of retrieval
sf_bp_geo |>
create_map(start_date = "20205-01-01", end_date = "2025-07-15")
}
}
