% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{create_map_input}
\alias{create_map_input}
\title{Prepare Address Data for Mapping}
\usage{
create_map_input(df_bp_geo)
}
\arguments{
\item{df_bp_geo}{A data frame containing project and address information,
including IDs, publication dates, deadlines, and components of the project
location address.}
}
\value{
A data frame with unique project entries and a cleaned, concatenated
address field.
}
\description{
Aggregates and formats geographic and descriptive information from a data
frame containing project-related geolocation and metadata for use in mapping
applications.
}
