% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_and_download_xml.R
\name{download_xml_files}
\alias{download_xml_files}
\title{Download Baupublikationen xml-files}
\usage{
download_xml_files(path, params)
}
\arguments{
\item{path}{A path to a folder where new and already downloaded Baupublikatonen
are stored}

\item{params}{A \code{list} with a selection of parameters that refers to the
publication in a given Amtsblatt (also the one of other cantons than Zurich).
For the parameters for the Baupaublikationen, see the example.
For all parameters, see the
\href{https://www.amtsblattportal.ch/docs/api/#_api_reference}{API-Documentation}
of the Amtsblattportal.}
}
\description{
Downloading available xml-files based on your input paramater consisting of
the meta information and content of all Baupublikationen.
}
\details{
Note that you can specify  \code{publicationDate.start} and \code{publicationDate.end}
in your "params" (also see example).
}
\examples{
\dontrun{
  params <- list(
  publicationStates = "PUBLISHED",
  rubrics = "BP-ZH",
  subrubrics = "BP-ZH01",
  publicationDate.start = "2025-06-01",
  publicationDate.end = as.character(Sys.Date()),
  cantons = "ZH",
  pageRequest.size = 100
  )
  
  path <- "YOUR_STORAGE_FOlDER/"
  
  download_xml_files(path, params)
}
}
