% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_to_df.R
\name{expand_df}
\alias{expand_df}
\title{Expand data frame}
\usage{
expand_df(df, index_var)
}
\arguments{
\item{df}{A data frame to be expanded}

\item{index_var}{A variable based on which the expansion should be conducted}
}
\value{
An expanded data frame
}
\description{
This function expands the data frame based on an index variable. In the input
data frame, the index variable is already expanded (e.g., location 1,
location 2, ...) while all other variables (here rows) only occur once. The
purpose of this function is therefore to expand all non-index variables by
the number of unique index values.
}
