% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_to_df.R
\name{extract_leaves}
\alias{extract_leaves}
\title{Extract leafs from xml and transform to data frame}
\usage{
extract_leaves(node, path = "", index_list = list())
}
\arguments{
\item{node}{An arbitrary node which should be flattend, for the
Baupublikationen-case it is the root node}

\item{path}{An empty path that is enriched by iterating through the nodes.
A single path is equivalent to variable name}

\item{index_list}{An empty list which is updated if there multiple nodes/leafs
with the same name}
}
\value{
A data frame in long format
}
\description{
Function to recursively extract all leaf nodes with their full paths.
}
