% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_url}
\alias{extract_url}
\title{Extract all urls from a single page during the API call}
\usage{
extract_url(page, base_url, params)
}
\arguments{
\item{page}{The current page number}

\item{base_url}{A character string, representing the base url for making the
API call.}

\item{params}{A \code{list} with a selection of parameters that refers to the
publication in a given Amtsblatt (also the one of other cantons than Zurich).
For the parameters for the Baupaublikationen, see the example.
For all parameters, see the
\href{https://www.amtsblattportal.ch/docs/api/#_api_reference}{API-Documentation}
of the Amtsblattportal.}
}
\value{
Returns a data frame containing a selection of meta information of
all Baupublikationen of the current page including the \code{id}.
}
\description{
This function creates a data frame based on the current page of the pagination
process contain a selection of a given Baupuplikation's meta data including
the required \code{id}.
}
