% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_refercing.R
\name{geo_info_check}
\alias{geo_info_check}
\title{Check if data was lost in geo-matching process}
\usage{
geo_info_check(df_bp, df_bp_geo)
}
\arguments{
\item{df_bp}{An expanded data frame in wide format consisting of publications}

\item{df_bp_geo}{A data frame consisting of publication including a polygon
column}
}
\description{
This functions checks if rows are lost in the geo-matching process and if so,
returns the lost IDs.
}
