% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_baupub_helpers.R
\name{get_columns_of_interest}
\alias{get_columns_of_interest}
\title{Define and select the columns of interest for the OGD data frame}
\usage{
get_columns_of_interest(df_bp)
}
\arguments{
\item{df_bp}{An expanded data frame in wide format consisting of publications}
}
\description{
Define and select the columns of interest for the OGD data frame
}
