% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liegenschaften_ZH.R
\name{get_geo_info}
\alias{get_geo_info}
\title{Get polygons for publications}
\usage{
get_geo_info(df_bp, sf_liegenschaften)
}
\arguments{
\item{df_bp}{An expanded data frame in wide format consisting of publications}

\item{sf_liegenschaften}{A shapefile containing the Liegenschaften-data
geometry and metadata.}
}
\value{
A data frame consisting of publication including a polygon column
}
\description{
By combining the BFS- and the cadaster-number (unique combination), this
function adds a polygon for each combination in a given publication - if
being found. Note that \code{clean_cadaster} cannot cover all possible cases.
Hence, so combinations cannot be matched. Also note that a single project can
involve multiple parcels.
}
