% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_refercing.R
\name{get_giszh_api_download_url}
\alias{get_giszh_api_download_url}
\title{Request Download URL from GISZH API}
\usage{
get_giszh_api_download_url(source_location, email_address)
}
\arguments{
\item{source_location}{An integer identifying the product ID to download
(e.g., 404 for Liegenschaften).}

\item{email_address}{A string containing a valid email address to receive the
data request.}
}
\value{
A string containing the download URL for the requested product.
}
\description{
Sends a request to the GISZH API to generate a download URL for a given
product. Waits until the data is ready and returns the download URL.
}
\examples{
\dontrun{
url <- get_giszh_api_download_url(404, "your.email@example.com")
}
}
