% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{get_legal_entity_info}
\alias{get_legal_entity_info}
\title{Get legal entity info for companies}
\usage{
get_legal_entity_info(df_bp)
}
\arguments{
\item{df_bp}{An expanded data frame in wide format consisting of publications}
}
\description{
This function adds the legal entitiy info for building contractor, project
framer and building contractor delegation companies.
}
