% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_liegenschaften_layer.R
\name{get_liegenschaften_layer}
\alias{get_liegenschaften_layer}
\title{Download and Load Liegenschaften Layer from GISZH API}
\usage{
get_liegenschaften_layer(file_destination, email_address, retrieval_day)
}
\arguments{
\item{file_destination}{A string specifying the destination folder where the
shapefile should be stored.}

\item{email_address}{A string containing a valid email address required by
the GISZH API for data download.}

\item{retrieval_day}{A string specifying the weekday (e.g., "Monday") on which
the data should be re-downloaded.}
}
\value{
An sf object containing the Liegenschaften spatial data
}
\description{
Downloads the Liegenschaften shapefile from the GISZH API if the specified
retrieval day matches the current weekday or if the shapefile does not already
exist locally. The function then reads the shapefile into an sf object.
}
\examples{
\dontrun{
get_liegenschaften_layer("data", "your.email@example.com", "Monday")
}
}
