% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_and_download_xml.R
\name{total_entries_check}
\alias{total_entries_check}
\title{Determine the maximal number of available Baupublikationen}
\usage{
total_entries_check(url, params)
}
\arguments{
\item{url}{Skeleton url which has to be enriched with parameters}

\item{params}{A \code{list} with a selection of parameters that refers to the
publication in a given Amtsblatt (also the one of other cantons than Zurich).
For the parameters for the Baupaublikationen, see the example.
For all parameters, see the
\href{https://www.amtsblattportal.ch/docs/api/#_api_reference}{API-Documentation}
of the Amtsblattportal.}
}
\value{
A numeric value of the number of the available Baupublikationen
}
\description{
Sends a GET request to a specified URL with query parameters, parses the XML
response, and extracts the total number of entries. If the number of entries
exceeds 9,999, an error is raised, prompting the user to select a narrower
time frame.
}
\details{
It is also possible to determine the number of publications for a certain
period by adding the parameter \code{publicationDate.start} and \code{publicationDate.end}.
If so, at least \code{publicationDate.end} in the following format "YYYY-MM-DD" is
required. For more information regarding the parameters see the
\href{https://www.amtsblattportal.ch/docs/api/#_core_concepts}{API-documentation}.
}
\examples{
\dontrun{
  params <- list(
  publicationStates = "PUBLISHED",
  rubrics = "BP-ZH",
  subrubrics = "BP-ZH01",
  publicationDate.start = "2025-06-06",
  publicationDate.end = as.character(Sys.Date()),
  cantons = "ZH",
  pageRequest.size = 100
  )
  
  url <- "https://amtsblattportal.ch/api/v1/publications/xml?"
  
  total_entries_check(url, params) 
}

}
