% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_new_pub_url}
\alias{get_new_pub_url}
\title{Retrieving the url of all newly published Baupublikationen}
\usage{
get_new_pub_url(page_size = 100, url, params, df_bp)
}
\arguments{
\item{page_size}{The number of entries per page through which it will be
iterrated. The smaller the number, the faster (hence default = 100).}

\item{url}{Skeleton url which has to be enriched with parameters}

\item{params}{A \code{list} with a selection of parameters that refers to the
publication in a given Amtsblatt (also the one of other cantons than Zurich).
For the parameters for the Baupaublikationen, see the example.
For all parameters, see the
\href{https://www.amtsblattportal.ch/docs/api/#_api_reference}{API-Documentation}
of the Amtsblattportal.}

\item{df_bp}{Existing OGD-csv-distribution}
}
\description{
This function retrieves a vector that consists the IDs through which a single
publicaton can be identified and its content can be downloaded.
}
